requirements.txt
attr==0.3.2
attrs==22.2.0
ConfigParser==5.3.0
Cython==0.29.33
dl==0.1.0
docutils==0.19
HTMLParser==0.0.2
ipython==8.11.0
ipywidgets==8.0.5
Jinja2==3.1.2
keyring==23.13.1
lockfile==0.12.2
numpy==1.24.2
Pillow==9.4.0
protobuf==3.20.3
pyOpenSSL==23.1.1
railroad==0.5.0
Sphinx==6.1.3
streamlit==1.20.0
tornado==6.2
scikit_learn==1.2.2